% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregate_study_info}
\alias{aggregate_study_info}
\title{create a table with aggregated data: each row contains information
about control and treatments of a single study}
\usage{
aggregate_study_info(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
data.frame
}
\description{
create a table with aggregated data: each row contains information
about control and treatments of a single study
}
