% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-stat-analysis.R
\name{animal_info_classification}
\alias{animal_info_classification}
\title{Generate table representing number of animals in classification groups}
\usage{
animal_info_classification(data)
}
\arguments{
\item{data}{final classification data}
}
\value{
data frame
}
\description{
Generate table representing number of animals in classification groups
}
