% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-stat-analysis.R
\name{assess_efficacy}
\alias{assess_efficacy}
\title{Credible interval (or say “Bayesian confidence interval”) of the mean
difference between two groups (treatment and reference) is used to assess the
efficacy. If 0 falls outside the interval, the drug was considered
significantly effective}
\usage{
assess_efficacy(data, reference = "Control")
}
\arguments{
\item{data}{prediction results}

\item{reference}{name of the reference treatment}
}
\value{
dataframe with information about drug efficacy
}
\description{
Credible interval (or say “Bayesian confidence interval”) of the mean
difference between two groups (treatment and reference) is used to assess the
efficacy. If 0 falls outside the interval, the drug was considered
significantly effective
}
