% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-exclude.R
\name{below_min_points}
\alias{below_min_points}
\title{makes df with data to be excluded}
\usage{
below_min_points(df, min_points)
}
\arguments{
\item{df}{initial data frame}

\item{min_points}{minimum number of data points for one animal_id per study}
}
\value{
df
}
\description{
makes df with data to be excluded
}
