% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-stat-analysis.R
\name{calc_probability}
\alias{calc_probability}
\title{Calculate probability of categories}
\usage{
calc_probability(data)
}
\arguments{
\item{data}{data frame with predictions}
}
\value{
data frame
}
\description{
Calculate probability of categories
}
