% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_survived}
\alias{calc_survived}
\title{Calculate percentage of survived animals}
\usage{
calc_survived(df)
}
\arguments{
\item{df}{data frame}
}
\value{
data frame
}
\description{
Calculate percentage of survived animals
}
