% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-model.R
\name{change_time_multi}
\alias{change_time_multi}
\title{Get an array with change_time for studies from the population-level effects,
multiple studies}
\usage{
change_time_multi(model)
}
\arguments{
\item{model}{an object of class brmsfit}
}
\value{
data frame
}
\description{
Get an array with change_time for studies from the population-level effects,
multiple studies
}
