% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-model.R
\name{change_time_single}
\alias{change_time_single}
\title{Get a change time from the population-level effects, single study}
\usage{
change_time_single(model)
}
\arguments{
\item{model}{an object of class brmsfit}
}
\value{
a numeric vector of length one
}
\description{
Get a change time from the population-level effects, single study
}
