% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-model.R
\name{classify_data_point}
\alias{classify_data_point}
\title{Classify individual data points as Responders or Non-responders}
\usage{
classify_data_point(df_newstudy, pred_newstudy)
}
\arguments{
\item{df_newstudy}{data from new study}

\item{pred_newstudy}{data frame with predictions}
}
\value{
data frame with "Responder"/"Non-responder"
for individual data points
}
\description{
Classify individual data points as Responders or Non-responders
}
