% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-stat-analysis.R
\name{classify_subcategories}
\alias{classify_subcategories}
\title{Make predictions for subcategories}
\usage{
classify_subcategories(data, model)
}
\arguments{
\item{data}{data frame with classification results}

\item{model}{object of class brmsfit}
}
\value{
data frame
}
\description{
Make predictions for subcategories
}
