% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-model.R
\name{classify_type_responder}
\alias{classify_type_responder}
\title{Classify tumour based on the growth rate and the p_value for a two-sided
T test
Tumour will be considered as "Non-responder", "Modest responder",
"Stable responder" or "Regressing responder"}
\usage{
classify_type_responder(df)
}
\arguments{
\item{df}{data frame}
}
\value{
data frame with a new column \code{classify_tumour}
}
\description{
Classify tumour based on the growth rate and the p_value for a two-sided
T test
Tumour will be considered as "Non-responder", "Modest responder",
"Stable responder" or "Regressing responder"
}
