% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_string}
\alias{clean_string}
\title{function to remove hyphens, underscores, spaces and transform to lowercase}
\usage{
clean_string(string)
}
\arguments{
\item{string}{to modify}
}
\value{
modified string
}
\description{
function to remove hyphens, underscores, spaces and transform to lowercase
}
