% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-model.R
\name{control_growth_plot}
\alias{control_growth_plot}
\title{Function to plot a control growth profile}
\usage{
control_growth_plot(df, model_type, col_palette)
}
\arguments{
\item{df}{data frame}

\item{model_type}{string}

\item{col_palette}{character palette}
}
\value{
ggplot object
}
\description{
Function to plot a control growth profile
}
