% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Tumour volume data over time for in-vivo studies}
\format{
A data frame with 1462 rows and 6 variables:
\describe{
\item{study}{study identifier}
\item{group}{group identifier}
\item{treatment}{treatment type}
\item{animal_id}{animal identifier}
\item{day}{day after implant}
\item{tumour_volume}{volume in mm3}
}
}
\usage{
example_data
}
\description{
A dataset containing the repeatedly measurements of tumour volume data over
time for individual animals.
}
\keyword{datasets}
