% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-exclude.R
\name{exclude_data}
\alias{exclude_data}
\title{Filter rows to exclude from the analysis}
\usage{
exclude_data(df, study_id_ex, animal_id_ex, day_ex, reason)
}
\arguments{
\item{df}{initial df}

\item{study_id_ex}{string: study id}

\item{animal_id_ex}{string: animal id}

\item{day_ex}{string: day}

\item{reason}{string: why it should be excluded}
}
\value{
dataframe with rows that meets exclusion criteria
}
\description{
Filter rows to exclude from the analysis
}
