% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand_palette}
\alias{expand_palette}
\title{Function to expand a vector of colors if needed}
\usage{
expand_palette(col_palette, n)
}
\arguments{
\item{col_palette}{character palette to color the treatments}

\item{n}{how many colors are needed}
}
\value{
a character vector of colors
}
\description{
Function to expand a vector of colors if needed
}
