% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-model.R
\name{f_start}
\alias{f_start}
\title{Calculate coefficients for a nonlinear model}
\usage{
f_start(df, x, y, r_change)
}
\arguments{
\item{df}{data frame with x as a predictor and y is an outcome}

\item{x}{predictor string}

\item{y}{outcome string}

\item{r_change}{numeric}
}
\value{
list of coefficients
}
\description{
Calculate coefficients for a nonlinear model
}
