% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-model.R
\name{get_responder}
\alias{get_responder}
\title{Classify tumour based on response status of individuals}
\usage{
get_responder(x, n)
}
\arguments{
\item{x}{character vector with response statuses of one animal}

\item{n}{consecutive measurements for classification}
}
\value{
"Responder" or "Non-responder"
}
\description{
Classify tumour based on response status of individuals
}
