% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{guess_match}
\alias{guess_match}
\title{function to search for the possible critical columns in a data.frame}
\usage{
guess_match(colnames_df, crit_cols)
}
\arguments{
\item{colnames_df}{a character vector with names}

\item{crit_cols}{a character vector}
}
\value{
list: possible match to each critical column
}
\description{
function to search for the possible critical columns in a data.frame
}
