% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-classification.R
\name{hide_outliers}
\alias{hide_outliers}
\title{Function to hide outliers in boxplots with jitterdodge
as suggested}
\usage{
hide_outliers(x)
}
\arguments{
\item{x}{plotly object}
}
\value{
plotly object without boxplot outliers
}
\description{
Function to hide outliers in boxplots with jitterdodge
as suggested
}
