% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_data}
\alias{load_data}
\title{function to read data from users (.csv or .xlsx files)}
\usage{
load_data(path, name)
}
\arguments{
\item{path}{path to a temp file}

\item{name}{filename provided by the web browser}
}
\value{
data frame
}
\description{
function to read data from users (.csv or .xlsx files)
}
