% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-model.R
\name{make_terms}
\alias{make_terms}
\title{Create a character vector with the names of terms from model, for which
predictions should be displayed
Specific values are specified in square brackets}
\usage{
make_terms(days, studies = NULL)
}
\arguments{
\item{days}{vector with days with which to predict}

\item{studies}{vector with studies with which to predict}
}
\value{
vector with values for predictions
}
\description{
Create a character vector with the names of terms from model, for which
predictions should be displayed
Specific values are specified in square brackets
}
