% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-model.R
\name{model_control}
\alias{model_control}
\title{Build model and make predictions}
\usage{
model_control(df_control, df_newstudy, method, end_day)
}
\arguments{
\item{df_control}{data frame with control data (including historical control,
if provided)}

\item{df_newstudy}{data frame, data from new study}

\item{method}{"Two-stage non-linear model" or "Linear model"}

\item{end_day}{period of time used for the statistical modelling of
the control data}
}
\value{
list: two data frames with prediction results
(for new study and for control data)
}
\description{
Build model and make predictions
}
