% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{notify_error_and_reset_input}
\alias{notify_error_and_reset_input}
\title{Display a popup message and reset fileInput}
\usage{
notify_error_and_reset_input(message_text)
}
\arguments{
\item{message_text}{the modal's text}
}
\description{
Display a popup message and reset fileInput
}
