% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-stat-analysis.R
\name{ordered_regression}
\alias{ordered_regression}
\title{Fit model (Bayesian ordered logistic regression)}
\usage{
ordered_regression(df, formula, n_cores)
}
\arguments{
\item{df}{data frame with classification results. Tumour classification is
converted into ordinal data}

\item{formula}{string}

\item{n_cores}{number of cores to use}
}
\value{
object of class brmsfit
}
\description{
Fit model (Bayesian ordered logistic regression)
}
