% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-stat-analysis.R
\name{plot_animal_info}
\alias{plot_animal_info}
\title{Plot representing number of animals in classification groups}
\usage{
plot_animal_info(data, col_palette)
}
\arguments{
\item{data}{final classification data}

\item{col_palette}{character palette}
}
\value{
ggplot object
}
\description{
Plot representing number of animals in classification groups
}
