% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-classification.R
\name{plot_class_gr}
\alias{plot_class_gr}
\title{Function to plot classification over growth rate}
\usage{
plot_class_gr(df, col_palette)
}
\arguments{
\item{df}{data frame}

\item{col_palette}{character palette}
}
\value{
ggplot object
}
\description{
Function to plot classification over growth rate
}
