% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-classification.R
\name{plot_class_tv}
\alias{plot_class_tv}
\title{Function to plot classification over tumour volume}
\usage{
plot_class_tv(df, col_palette, title_name)
}
\arguments{
\item{df}{data frame}

\item{col_palette}{named vector}

\item{title_name}{character}
}
\value{
ggplot object
}
\description{
Function to plot classification over tumour volume
}
