% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-stat-analysis.R
\name{plot_proportions}
\alias{plot_proportions}
\title{Plot estimated proportions}
\usage{
plot_proportions(data, col_palette)
}
\arguments{
\item{data}{table of the category prediction}

\item{col_palette}{character palette}
}
\description{
Plot estimated proportions
}
