% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-classification.R
\name{plot_waterfall}
\alias{plot_waterfall}
\title{Function to plot waterfall}
\usage{
plot_waterfall(df, col_palette, study_name)
}
\arguments{
\item{df}{data frame}

\item{col_palette}{character palette}

\item{study_name}{string: to show on title}
}
\value{
ggplot object
}
\description{
Function to plot waterfall
}
