% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotly_volume}
\alias{plotly_volume}
\title{Create volume plot for one-batch data}
\usage{
plotly_volume(
  df,
  col_palette = NULL,
  faceting_var,
  y_name,
  y_var,
  p_title,
  ...
)
}
\arguments{
\item{df}{data.frame, single-batch long format}

\item{col_palette}{character palette to color the treatments}

\item{faceting_var}{string}

\item{y_name}{string}

\item{y_var}{string: column name for y axis}

\item{p_title}{plot title}

\item{...}{arguments passed to plot_ly}
}
\value{
plotly object
}
\description{
Create volume plot for one-batch data
}
