% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-model.R
\name{predict_lm}
\alias{predict_lm}
\title{Make predictions, linear model}
\usage{
predict_lm(model, newdata, single)
}
\arguments{
\item{model}{a model object}

\item{newdata}{data frame in which to look for variables with which to
predict}

\item{single}{logical: TRUE if single study experiment}
}
\value{
data frame with predictions
}
\description{
Make predictions, linear model
}
