% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-model.R
\name{predict_nlm_multi}
\alias{predict_nlm_multi}
\title{Make predictions based on non-linear model, multiple studies}
\usage{
predict_nlm_multi(model, newdata, change_time)
}
\arguments{
\item{model}{an object of class brmsfit}

\item{newdata}{data frame in which to look for variables
with which to predict}

\item{change_time}{data frame}
}
\value{
data frame with predictions
}
\description{
Make predictions based on non-linear model, multiple studies
}
