% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-model.R
\name{predict_nlm_single}
\alias{predict_nlm_single}
\title{Make predictions based on non-linear model, single study}
\usage{
predict_nlm_single(model, newdata, change_time)
}
\arguments{
\item{model}{an object of class brmsfit}

\item{newdata}{data frame in which to look for variables
with which to predict}

\item{change_time}{numeric}
}
\value{
data frame with predictions
}
\description{
Make predictions based on non-linear model, single study
}
