% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-stat-analysis.R
\name{predict_regr_model}
\alias{predict_regr_model}
\title{Make predictions}
\usage{
predict_regr_model(model, df)
}
\arguments{
\item{model}{object of class brmsfit}

\item{df}{data frame with classification results}
}
\value{
data frame
}
\description{
Make predictions
}
