% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(...)
}
\arguments{
\item{...}{additional options passed to shinyApp()}
}
\value{
No return value, called for the shiny app interface
}
\description{
Run the Shiny Application
}
