% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-model.R
\name{run_nl_model}
\alias{run_nl_model}
\title{Fit nonlinear model - continuous hinge function}
\usage{
run_nl_model(start, df_mod, formula, n_cores)
}
\arguments{
\item{start}{df with coefficients}

\item{df_mod}{data of all variables used in the model}

\item{formula}{an object of class brmsformula}

\item{n_cores}{number of cores to use}
}
\value{
object of class brmsfit
}
\description{
Fit nonlinear model - continuous hinge function
}
