% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_waiter}
\alias{set_waiter}
\title{Set up a waiting screen}
\usage{
set_waiter(header)
}
\arguments{
\item{header}{text to display on loading screen}
}
\value{
object of a class waiter
}
\description{
Set up a waiting screen
}
